requirements.txt
appnope==0.1.0
decorator==4.0.6
dj-database-url==0.3.0
Django==1.8
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
gnureadline==6.3.3
gunicorn==19.4.5
httpie==0.9.3
ipython==4.0.1
ipython-genutils==0.1.0
Markdown==2.6.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
pyparsing==2.0.7
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wheel==0.24.0
